t/functional: create a functional test suite
We've been using the microsoft/scalar repository to run "functional
tests" that check that microsoft/git works against a real instance of
Azure DevOps. Start creating those kinds of tests in microsoft/git.

The first test is a simple one: check that 'scalar diagnose' works. This
could probably be done without a functional test, since we have tests
that check the GVFS Protocol using a test helper. The big thing we need
is a test that checks that the shared object cache is present so we can
prevent bugs like #541.

Since the test script starts with a 'scalar clone', this test also
implicitly tests that those clones work.

Signed-off-by: Derrick Stolee <derrickstolee@github.com>