t921*: test scalar behavior starting maintenance
A user recently reported issues with 'scalar register' and 'scalar
clone' in that they failed when the system had permissions locked down
so both 'crontab' and 'systemctl' commands failed when trying to enable
background maintenance.

This hard error is undesirable, but let's create tests that demonstrate
this behavior before modiying the behavior. We can use
GIT_TEST_MAINT_SCHEDULER to guarantee failure and check the exit code
and error message. This must be set in a sub-shell to avoid issues with
check-non-portable-shell complaining about the environment variable
acting around a shell function.

Signed-off-by: Derrick Stolee <derrickstolee@github.com>