t0450: generalize to non-builtins
The test script t0450-txt-doc-vs-help.sh is currently limited to
builtins, due to operating on the output of 'git --list-cmds=builtins'
and no other possible commands. For example, 'git-remote-https' is not
technically a builtin and must be executed in dashed form. The 'scalar'
command also does not work with this setup.

Make the necessary changes to have the identical test coverage but
instead refer to builtins in their spaced form: "git <builtin>".

Using two helpers, cmd_to_dashed and cmd_to_underscore, are helpful for
finding the Documentation/git-<builtin>.txt file and defining the
BUILTIN_TXT_git_<builtin> prerequisites.

The format of t/t0450/txt-help-mismatches needs updating to work with
this more general form, as well.

Signed-off-by: Derrick Stolee <derrickstolee@github.com>