t0450: add 'scalar' as a possible help text test
The previous change created the possibility for checking that the
'scalar -h' output matches the synopsis from Documentation/scalar.txt.
It currently does _not_ match, so we need to add it to the mismatch
file.

Use a here-doc in case we want to test other commands that do not appear
in the 'git --list-cmds=builtins' output. It will be easy to add new
lines to the end of the 'cmds' file.

Signed-off-by: Derrick Stolee <derrickstolee@github.com>