scalar: usage matches help text
The previous change included the 'scalar' binary in the list of commands
to automatically verify that its '-h' helptext matches its documented
synopsis. However, the command was disabled as expected failures because
of the mismatch in the usage.

Resolve the mismatched usage by also updating the argument parsing to
use OPT_SUBCOMMAND() for all of the included subcommands. This somewhat
simplifies the code as well as demonstrates a single place where the
usage needs to be updated.

This use of macros is modeled after the subcommand parsing in
builtin/commit-graph.c.

Note that 'scalar help' and 'scalar version' are not included in the
documented synopsis, so they are omitted from the usage string. However,
I did notice that 'scalar version' had a typo in its usage which is
resolved here ("verbose" to "version").

Signed-off-by: Derrick Stolee <derrickstolee@github.com>