post-command-hook: do not check environment
The post-command hook can be run during the atexit() handler, and it
checks first to see if the pre-command hook ran and then checks the
COMMAND_HOOK_LOCK environment variable to see if we should skip the
hook (to avoid recursion).

When running in the atexit() handler, this check for the environment
causes an allocation that could be problematic when mimalloc is shutting
down already.

Signed-off-by: Derrick Stolee <derrickstolee@github.com>