repository: create disable_replace_refs()
Several builtins depend on being able to disable the replace references
so we actually operate on each object individually. These currently do
so by directly mutating the 'read_replace_refs' global.

A future change will move this global into a different place, so it will
be necessary to change all of these lines. However, we can simplify that
transition by abstracting the purpose of these global assignments with a
method call.

The method call is scoped to a 'struct repository' pointer because we
will move the state of reading replace references into the repository
settings struct.

Signed-off-by: Derrick Stolee <derrickstolee@github.com>