repository: move read_replace_refs into settings
The 'read_replace_refs' global specifies whether or not we should
respect the references of the form 'refs/replace/<oid>' to replace which
object we look up when asking for '<oid>'. This global has caused issues
when it is not initialized properly, such as in b6551feadfd (merge-tree:
load default git config, 2023-05-10).

To make this more robust, move the global into the repository settings
struct. In addition to making this repository-scoped, we can also use
prepare_repo_settings() as a guarantee that it has been initialized
before we consume the value. This will prevent future repeats of the
issue where we leave a bug due to not loading default config soon
enough.

One careful thing that must be done is to modify disable_replace_refs()
to only call prepare_repo_settings() if the repository actually has a
Git dir. Without one, we will hit a BUG() statement, but also there is
no reason to actually focus on replace refs (because there are no refs
or objects to care about). This behavior is verified by several tests
across the test suite.

Signed-off-by: Derrick Stolee <derrickstolee@github.com>