config: prepare before using trust_ctime, check_stat
The prepare_default_config() method is a wrapper that checks if we have
loaded default config, and loads it if we have not. Protect uses of
'trust_ctime' and 'check_stat' with these calls to make sure the
variable has been initialized before use.

Both of these variables are used multiple times in match_stat_data(), so
all uses of both variables within that method are protected by this
single call.

Signed-off-by: Derrick Stolee <derrickstolee@github.com>