config: prepare before using is_bare_repository_cfg
The prepare_default_config() method is a wrapper that checks if we have
loaded default config, and loads it if we have not. Protect uses of
'is_bare_repository_cfg' with these calls to make sure the variable has
been initialized before use.

In the case of cmd_init_db(), the variable is set by the '--bare'
command-line option. Be sure to load the default config before parsing
options (and using the values multiple times in that method) so we do
not accidentally un-set the value through a later use of
prepare_default_config().

In setup.c, there are several uses of is_bare_repository_cfg, but these
cannot be prefaced with prepare_default_config(). The reasoning is that
these are used while setting up the repository at the beginning based on
the file conditions. Loading config at this time is inappropriate and
leads to test failures, such as tests in t0007-git-var.sh. Perhaps a
longer-term approach would be to split the variable into two: the
"guess" at the bareness of the repository, and then the configured
value. For now, we have isolated the variable somewhat but leave that
work to #leftoverbits.

Signed-off-by: Derrick Stolee <derrickstolee@github.com>