config: prepare before using is_bare_repository_cfg
The prepare_default_config() method is a wrapper that checks if we have
loaded default config, and loads it if we have not. Protect uses of
'is_bare_repository_cfg' with these calls to make sure the variable has
been initialized before use.

In the case of cmd_init_db(), the variable is set by the '--bare'
command-line option. Be sure to load the default config before parsing
options (and using the values multiple times in that method) so we do
not accidentally un-set the value through a later use of
prepare_default_config().

Signed-off-by: Derrick Stolee <derrickstolee@github.com>