setup: use strcmp during early config
The safe_directory_cb method was introduced in 8959555cee7
(setup_git_directory(): add an owner check for the top-level directory,
2022-03-02) as part of iterating over system and global config very
early in the Git process.

One of the things this does is compare filesystem paths using
fspathcmp(). The implementation of that method toggles between
strcasecmp() and strcmp() depending on whether the global 'ignore_case'
is set. However, this variable is set during the load of our default
config, so has not been initialized at this point. Thus, this is
actually always using strcmp() at this point.

An upcoming change will introduce a check that the default config has
been loaded before accessing the 'ignore_case' variable, but loading
that config in the middle of our safe early config causes problems due
to overlapping config parsing. Avoid this issue by using strcmp()
directly.

Signed-off-by: Derrick Stolee <derrickstolee@github.com>