fixup! scalar: configure maintenance during 'reconfigure'
This microsoft/git-specific change went through a few refactors to get
into this state, but now it's sub-optimal: it "succeeds" if setting the
config -or- toggling maintenance succeeds. The logic used to be that it
set a failure in either of these modes, but the variable was changed to
be a success.

Merge these if statements together. This should simplify the logic for
now. The resulting squashed commit could also be upstreamed.

Signed-off-by: Derrick Stolee <stolee@gmail.com>