checkout: ignore skip-worktree when virtual
As documented in microsoft/vfsforgit#1812, attempting to restore a file
fails when using either of these commands in a VFS for Git repo:

  git restore --source=<commit> <path>
  git checkout <commit> -- <path>

To discover the issue, I debugged such a call and found that since
opt->ignore_skipworktree is not set, that the restore will not update
the index when the file is not hydrated.

I verified that this works as expected, including that the file on-disk
is projecting the new index version.

Signed-off-by: Derrick Stolee <stolee@gmail.com>