fixup! Adjust README.md for microsoft/git
Adjust the currently-broken GVFS Protocol link.

The current link points to a page that has gone away. While there is
https://web.archive.org/web/20210302002834/https://docs.microsoft.com/en-us/azure/devops/learn/git/gvfs-architecture#gvfs-protocol
that _could_ be used to reinstate the link, it is not actually the best
document to which to point the keen reader. We already point interested
parties to the VFSforGit repository's documentation elsewhere, so let's
also do that in the README.

This fixes https://github.com/microsoft/git/issues/628.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>