fixup! github: use federated auth for Azure login
One "funny" quirk of GitHub Actions is that some of their YAML
constructs that users _assume_ to be cumulative, are not. One example is
the `permissions` block: Most readers will assume that the following
YAML will combine `contents` and `id-token` permissions:

	permissions:
	  id-token: write

	jobs:
	  my-job:
	    permissions:
	      contents: write

However, that is not the case! The inner `permissions` block completely
negates the outer `permissions` block.

So let's just repeat ourselves to force GitHub Actions to understand
which permissions we need.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>