Revert "midx-write.c: use `--stdin-packs` when repacking"
This reverts commit b7d6f23a17110d597d58f4a8e1b34b7a72c43fe1.

This commit caused a regression in the incremental-repack maintenance
step where not all objects in a pack-file are repacked, thus 'git
multi-pack-index expire' cannot delete the pack-files. This will require
some careful explanation upstream and some careful untangling of how
server-side repacks may be depending on the new logic. Since
microsoft/git only serves client users, it is safe to revert this in
isolation for now.

Signed-off-by: Derrick Stolee <stolee@gmail.com>