backfill: basic functionality and tests
The default behavior of 'git backfill' is to fetch all missing blobs that
are reachable from HEAD. Document and test this behavior.

The implementation is a very simple use of the path-walk API, initializing
the revision walk at HEAD to start the path-walk from all reachable commits.
Ignore the object arrays that correspond to tree entries.

Signed-off-by: Derrick Stolee <stolee@gmail.com>